newsimages25/workflows/Headline_Hunters_RETRIEVAL/requirements.txt
torch
transformers
requests
beautifulsoup4
pillow
numpy
scikit-learn
tqdm
faiss-cpu
lxml